[Feature] 파티 내용 조회 기능 구현

+++ src/main/java/meltingpot/server/domain/repository/party/PartyRepository.java
@@ -3,7 +3,6 @@ package meltingpot.server.domain.repository.party;
 import meltingpot.server.domain.entity.party.Party;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-public interface PartyRepository extends JpaRepository<Party, Long> {
+public interface PartyRepository extends JpaRepository<Party, Integer> {
     Party findByChatRoomId(Long chatRoomId);
 }
